requirements.txt
pyyaml
grr-api-client
pygelf
requests